chore: Remove GptSession utils
