+++ cpp/tensorrt_llm/batch_manager/rnnStateManager.cpp
@@ -17,7 +17,7 @@
 
 #include "tensorrt_llm/batch_manager/rnnStateManager.h"
 #include "tensorrt_llm/common/assert.h"
-#include "tensorrt_llm/runtime/utils/sessionUtils.h"
+#include "tensorrt_llm/runtime/utils/runtimeUtils.h"
 
 using namespace tensorrt_llm::runtime;
