+++ cpp/tensorrt_llm/runtime/utils/runtimeUtils.h
@@ -1,0 +1,44 @@
+/*
+ * Copyright (c) 2022-2024, NVIDIA CORPORATION.  All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#pragma once
+
+#include "tensorrt_llm/runtime/modelConfig.h"
+#include "tensorrt_llm/runtime/worldConfig.h"
+
+#include <cstdint>
+#include <string>
+#include <vector>
+
+namespace tensorrt_llm::runtime
+{
+class TllmRuntime;
+
+namespace utils
+{
+
+int initDevice(WorldConfig const& worldConfig);
+
+std::vector<uint8_t> loadEngine(std::string const& enginePath);
+
+void insertTensorVector(StringPtrMap<ITensor>& map, std::string const& key, std::vector<ITensor::SharedPtr> const& vec,
+    SizeType32 indexOffset, std::vector<ModelConfig::LayerType> const& layerTypes, ModelConfig::LayerType type);
+
+void insertTensorSlices(
+    StringPtrMap<ITensor>& map, std::string const& key, ITensor::SharedPtr const& tensor, SizeType32 indexOffset);
+
+} // namespace utils
+} // namespace tensorrt_llm::runtime
