+++ cpp/tests/batch_manager/trtEncoderModelTest.cpp
@@ -26,7 +26,7 @@
 #include "tensorrt_llm/runtime/rawEngine.h"
 #include "tensorrt_llm/runtime/tllmLogger.h"
 #include "tensorrt_llm/runtime/utils/numpyUtils.h"
-#include "tensorrt_llm/runtime/utils/sessionUtils.h"
+#include "tensorrt_llm/runtime/utils/runtimeUtils.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
